+++ foxes/models/turbine_models/__init__.py
@@ -7,3 +7,4 @@ from .sector_management import SectorManagement
 from .table_factors import TableFactors
 from .yaw2yawm import YAW2YAWM
 from .yawm2yaw import YAWM2YAW
+from .calculator import Calculator